requirements.txt
Django==1.4.3
PIL==1.1.7
boto==2.9.7
django-wysiwyg-redactor==0.3.5
wsgiref==0.1.2
